Foresight

Documentation for Foresight.

Foresight.PolarAxisTransformation

Foresight.adjustcam!

Foresight.axiscolorbar

Foresight.brighten

Foresight.clip

Foresight.darken

Foresight.demofigure

Foresight.foresight

Foresight.freeze!

Foresight.importall

Foresight.lscientific

Foresight.scientific

Foresight.seethrough

Foresight.widen

Foresight.@default_theme!

a = 1
b = 2
a+b

3

Foresight.PolarAxisTransformation — Type

PolarAxisTransformation(θ_0::Float64, direction::Int)

This struct defines a general polar-to-cartesian transformation, i.e.,

\[(r, θ) \to (r \cos(direction * (θ + θ_0)), r \sin(direction * (θ + θ_0)))\]

where θ is assumed to be in radians.

direction should be either -1 or +1, and θ_0 may be any value.

Foresight.adjustcam! — Function

Adjust the axis's scene's camera to conform to the given r-limits

Foresight.axiscolorbar — Method

axiscolorbar(ax, args...; kwargs...)

Create a colorbar for the given ax axis. The args argument is passed to the Colorbar constructor, and the kwargs argument is passed to the Colorbar constructor as keyword arguments. The position argument specifies the position of the colorbar relative to the axis, and can be one of :rt (right-top), :rb (right-bottom), :lt (left-top), :lb (left-bottom). The default value is :rt.

Example

f = Figure()
ax = Axis(f[1, 1])
x = -5:0.01:5
plot!(ax, x, x->sinc(x), color=1:length(x), colormap=sunset)
axiscolorbar(ax; label="Time (m)")

Foresight.brighten — Method

brighten(c::T, β)

Brighten a color c by a factor of β by blending it with white. β should be between 0 and 1.

Example

brighten(cornflowerblue, 0.5)

Foresight.clip — Function

tmpfile = clip(fig=Makie.current_figure(), fmt=:png; kwargs...)

Save the current figure to a temporary file and copy it to the clipboard. kwargs are passed to Makie.save.

Example

f = plot(-5:0.01:5, x->sinc(x))
clip(f)

Foresight.darken — Method

darken(c::T, β)

Darken a color c by a factor of β by blending it with black. β should be between 0 and 1.

Example

darken(cornflowerblue, 0.5)

Foresight.demofigure — Method

demofigure()

Produce a figure showcasing the current theme.

Foresight.foresight — Method

foresight(options...; font=foresightfont())

Return the default Foresight theme. The options argument can be used to modify the default values, by passing keyword arguments with the names of the attributes to be changed and their new values.

Some vailable options are:

:dark: Use a dark background and light text.

:transparent: Make the background transparent.

:minorgrid: Show minor gridlines.

:serif: Use a serif font.

:redblue: Use a red-blue colormap.

:gray: Use a grayscale colormap.

Foresight.freeze! — Method

freeze!(ax::Union{Axis, Axis3, Figure}=current_figure())

Freeze the limits of an Axis or Axis3 object at their current values.

Example

ax = Axis();
plot!(ax, -5:0.01:5, x->sinc(x))
freeze!(ax)

Foresight.importall — Method

importall(module)

Return an array of expressions that can be used to import all names from a module.

Example

importall(module) .|> eval

Foresight.lscientific — Function

lscientific(x::Real, sigdigits=2)

Return a string representation of a number in scientific notation with a specified number of significant digits. This is not an L-string.

Example

lscientific(1/123.456, 3) # "8.10 \times 10^{-3}"

Foresight.scientific — Function

scientific(x::Real, sigdigits=2)

Return a string representation of a number in scientific notation with a specified number of significant digits.

Arguments

x::Real: The number to be formatted.

sigdigits::Int=2: The number of significant digits to display.

Returns

A string representation of the number in scientific notation with the specified number of significant digits.

Example

scientific(1/123.456, 3) # "8.10 × 10⁻³"

Foresight.seethrough — Function

seethrough(C::ContinuousColorGradient, start=0.5, stop=1.0)

Convert a color gradient into a transparent version

Examples

C = sunrise;
transparent_gradient = seethrough(C)

Foresight.widen — Function

Slightly widen an interval by a fraction δ

Foresight.@default_theme! — Macro

default_theme!(thm)

Set the default theme to thm and save it as a preference. The change will take effect after restarting Julia.

Example

default_theme!(foresight())